refact: reuse character item widget for both pages: characters list and characters details pages
