+++ apps/mottu_marvel/lib/modules/characters/presentation/pages/character_details_page.dart
@@ -6,6 +6,7 @@ import 'package:mottu_design_system/mottu_design_system.dart';
 
 import '../../data/models/marvel_response_model.dart';
 import '../router/character_router.dart';
+import '../widgets/character_item.dart';
 import '../widgets/characters_page_progress_indicator.dart';
 
 class CharacterDetailsPage extends StatefulWidget {
@@ -71,7 +72,7 @@ class _CharacterDetailsPageState extends State<CharacterDetailsPage> {
             final List<Widget> charactersList = controller.charactersList
                 .map((eachCharacter) => Padding(
                       padding: const EdgeInsets.symmetric(horizontal: 8.0, vertical: 20.0),
-                      child: _MarvelCharacterItem(
+                      child: MarvelCharacterItem(
                         marvelCharacter: eachCharacter,
                       ),
                     ))
@@ -93,35 +94,6 @@ class _CharacterDetailsPageState extends State<CharacterDetailsPage> {
                 );
         }),
       ],
-    );
-  }
-}
-
-class _MarvelCharacterItem extends StatelessWidget {
-  const _MarvelCharacterItem({super.key, required this.marvelCharacter});
-
-  final MarvelCharacter marvelCharacter;
-
-  @override
-  Widget build(BuildContext context) {
-    return InkWell(
-      onTap: () => CharacterRouter.goToCharacterDetails(marvelCharacter),
-      child: Row(
-        children: [
-          MottuNetworkImage(
-            url: marvelCharacter.imageUrl,
-          ),
-          SizedBox(
-            width: 20,
-          ),
-          Text(
-            marvelCharacter.name,
-            style: TextStyle(
-              color: Colors.white,
-            ),
-          )
-        ],
-      ),
     );
   }
 }
